add a color picker to applications features
