+++ applications/kocolor/features/cosmetics/src/main/java/com/zoewave/probase/kocolor/features/cosmetics/ui/CosmeticsScreen.kt
@@ -1,10 +1,8 @@
 package com.zoewave.probase.kocolor.features.cosmetics.ui
 
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -53,7 +51,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Slider
 import androidx.compose.material3.SliderDefaults
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -62,7 +59,6 @@ import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableStateMapOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -70,12 +66,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
@@ -83,6 +74,10 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.AsyncImage
+import com.zoewave.probase.features.graphics.colorpicker.ui.ColorPickerDialog
+import com.zoewave.probase.features.graphics.colorpicker.util.isColorDark
+import com.zoewave.probase.features.graphics.colorpicker.util.parseColor
+import com.zoewave.probase.features.graphics.colorpicker.util.toHex
 import com.zoewave.probase.kocolor.model.CosmeticCategory
 import com.zoewave.probase.kocolor.model.CosmeticItem
 import com.zoewave.probase.kocolor.model.KoColorRoute
@@ -582,210 +577,6 @@ fun CosmeticProductCard(
     }
 }
 
-fun isColorDark(color: Color): Boolean {
-    val luminance = 0.299 * color.red + 0.587 * color.green + 0.114 * color.blue
-    return luminance < 0.5
-}
-
-@Composable
-fun FullColorPicker(
-    uiState: String,
-    onEvent: (String) -> Unit,
-    navTo: (KoColorRoute) -> Unit
-) {
-    val controller = rememberColorPickerController(initialColor = parseColor(uiState))
-
-    AlertDialog(
-        onDismissRequest = { navTo(KoColorRoute.Back) },
-        title = { Text("Pick Custom Color", fontWeight = FontWeight.Bold) },
-        text = {
-            Column(
-                modifier = Modifier.padding(vertical = 8.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                // Top Preview Tile
-                AlphaTile(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(60.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .border(1.dp, MaterialTheme.colorScheme.outline, RoundedCornerShape(12.dp)),
-                    controller = controller
-                )
-
-                // The Main HSV Color Wheel/Picker
-                HsvColorPicker(
-                    modifier = Modifier
-                        .size(240.dp)
-                        .padding(10.dp),
-                    controller = controller
-                )
-
-                // Brightness Slider
-                Column {
-                    Text("Brightness", style = MaterialTheme.typography.labelSmall)
-                    BrightnessSlider(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(35.dp),
-                        controller = controller
-                    )
-                }
-
-                Text(
-                    text = String.format("#%06X", (0xFFFFFF and controller.selectedColor.value.toArgb())),
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-        },
-        confirmButton = {
-            Button(onClick = { 
-                val hex = String.format("#%06X", (0xFFFFFF and controller.selectedColor.value.toArgb()))
-                onEvent(hex)
-                navTo(KoColorRoute.Back)
-            }) {
-                Text("Select Color")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = { navTo(KoColorRoute.Back) }) { Text("Cancel") }
-        }
-    )
-}
-
-class ColorPickerController(initialColor: Color) {
-    var selectedColor = mutableStateOf(initialColor)
-    
-    // Internal HSV state for the picker logic
-    var hue = mutableFloatStateOf(0f)
-    var saturation = mutableFloatStateOf(0f)
-    var value = mutableFloatStateOf(1f)
-
-    init {
-        val hsv = FloatArray(3)
-        android.graphics.Color.colorToHSV(
-            android.graphics.Color.rgb(
-                (initialColor.red * 255).toInt(),
-                (initialColor.green * 255).toInt(),
-                (initialColor.blue * 255).toInt()
-            ),
-            hsv
-        )
-        hue.floatValue = hsv[0]
-        saturation.floatValue = hsv[1]
-        value.floatValue = hsv[2]
-    }
-
-    fun updateFromHsv() {
-        val argb = android.graphics.Color.HSVToColor(floatArrayOf(hue.floatValue, saturation.floatValue, value.floatValue))
-        selectedColor.value = Color(argb)
-    }
-}
-
-@Composable
-fun rememberColorPickerController(initialColor: Color): ColorPickerController {
-    return remember(initialColor) { ColorPickerController(initialColor) }
-}
-
-@Composable
-fun AlphaTile(
-    modifier: Modifier,
-    controller: ColorPickerController
-) {
-    Box(
-        modifier = modifier.background(controller.selectedColor.value)
-    )
-}
-
-@Composable
-fun HsvColorPicker(
-    modifier: Modifier,
-    controller: ColorPickerController
-) {
-    Canvas(
-        modifier = modifier
-            .pointerInput(Unit) {
-                detectDragGestures { change, _ ->
-                    val w = size.width.toFloat()
-                    val h = size.height.toFloat()
-                    val center = Offset(w / 2f, h / 2f)
-                    val pos = change.position
-                    
-                    val dx = pos.x - center.x
-                    val dy = pos.y - center.y
-                    val radius = w / 2f
-                    
-                    val dist = Math.sqrt((dx * dx + dy * dy).toDouble()).toFloat()
-                    val angle = Math.toDegrees(Math.atan2(dy.toDouble(), dx.toDouble())).toFloat()
-                    
-                    controller.hue.floatValue = (if (angle < 0) angle + 360 else angle)
-                    controller.saturation.floatValue = (dist / radius).coerceIn(0f, 1f)
-                    controller.updateFromHsv()
-                }
-            }
-    ) {
-        val w = size.width
-        val h = size.height
-        val radius = w / 2f
-        val center = Offset(w / 2f, h / 2f)
-        
-        // Draw the color wheel background (simple radial sweep)
-        for (i in 0 until 360) {
-            val angle = Math.toRadians(i.toDouble()).toFloat()
-            val start = center
-            val end = Offset(
-                center.x + Math.cos(angle.toDouble()).toFloat() * radius,
-                center.y + Math.sin(angle.toDouble()).toFloat() * radius
-            )
-            val hsvColor = Color(android.graphics.Color.HSVToColor(floatArrayOf(i.toFloat(), 1f, 1f)))
-            
-            drawLine(
-                brush = Brush.linearGradient(listOf(Color.White, hsvColor), start, end),
-                start = start,
-                end = end,
-                strokeWidth = 2f
-            )
-        }
-        
-        // Draw the indicator
-        val indicatorAngle = Math.toRadians(controller.hue.floatValue.toDouble()).toFloat()
-        val indicatorRadius = controller.saturation.floatValue * radius
-        val indicatorOffset = Offset(
-            center.x + Math.cos(indicatorAngle.toDouble()).toFloat() * indicatorRadius,
-            center.y + Math.sin(indicatorAngle.toDouble()).toFloat() * indicatorRadius
-        )
-        
-        drawCircle(
-            color = if (controller.value.floatValue > 0.5f) Color.Black else Color.White,
-            radius = 8.dp.toPx(),
-            center = indicatorOffset,
-            style = Stroke(width = 2.dp.toPx())
-        )
-    }
-}
-
-@Composable
-fun BrightnessSlider(
-    modifier: Modifier,
-    controller: ColorPickerController
-) {
-    Slider(
-        value = controller.value.floatValue,
-        onValueChange = { 
-            controller.value.floatValue = it
-            controller.updateFromHsv()
-        },
-        valueRange = 0f..1f,
-        modifier = modifier,
-        colors = SliderDefaults.colors(
-            thumbColor = controller.selectedColor.value,
-            activeTrackColor = controller.selectedColor.value.copy(alpha = 0.5f)
-        )
-    )
-}
-
 @Composable
 fun EditCosmeticDialog(
     uiState: CosmeticItem,
@@ -802,10 +593,11 @@ fun EditCosmeticDialog(
     var showColorPicker by remember { mutableStateOf(false) }
 
     if (showColorPicker) {
-        FullColorPicker(
-            uiState = colorHex,
-            onEvent = { colorHex = it },
-            navTo = { showColorPicker = false }
+        ColorPickerDialog(
+            initialColor = parseColor(colorHex),
+            onColorSelected = { colorHex = it.toHex() },
+            onDismissRequest = { showColorPicker = false },
+            title = "Pick Product Color"
         )
     }
 
@@ -920,7 +712,7 @@ fun EditCosmeticDialog(
                                 shape = RoundedCornerShape(8.dp)
                             ),
                         color = try { 
-                            if (colorHex.isNotEmpty()) Color(android.graphics.Color.parseColor(colorHex)) else Color.Transparent 
+                            if (colorHex.isNotEmpty()) parseColor(colorHex) else Color.Transparent 
                         } catch (e: Exception) { Color.Transparent }
                     ) {
                         if (colorHex.isEmpty()) {
@@ -1074,10 +866,11 @@ fun AddCosmeticDialog(
     }
 
     if (showColorPicker) {
-        FullColorPicker(
-            uiState = colorHex,
-            onEvent = { colorHex = it },
-            navTo = { showColorPicker = false }
+        ColorPickerDialog(
+            initialColor = parseColor(colorHex),
+            onColorSelected = { colorHex = it.toHex() },
+            onDismissRequest = { showColorPicker = false },
+            title = "Pick Product Color"
         )
     }
 
@@ -1207,7 +1000,7 @@ fun AddCosmeticDialog(
                                 shape = RoundedCornerShape(8.dp)
                             ),
                         color = try { 
-                            if (colorHex.isNotEmpty()) Color(android.graphics.Color.parseColor(colorHex)) else Color.Transparent 
+                            if (colorHex.isNotEmpty()) parseColor(colorHex) else Color.Transparent
                         } catch (e: Exception) { Color.Transparent }
                     ) {
                         if (colorHex.isEmpty()) {
@@ -1253,14 +1046,6 @@ fun AddCosmeticDialog(
             TextButton(onClick = { navTo(KoColorRoute.Back) }) { Text("Cancel") }
         }
     )
-}
-
-fun parseColor(hex: String): Color {
-    return try {
-        Color(android.graphics.Color.parseColor(hex))
-    } catch (e: Exception) {
-        Color.Gray
-    }
 }
 
 @Preview(showBackground = true)
@@ -1366,18 +1151,6 @@ private fun SubCategoryCardPreview() {
     MaterialTheme {
         SubCategoryCard(
             uiState = Triple(CosmeticCategory.FOUNDATION, 3, false),
-            onEvent = {},
-            navTo = {}
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun FullColorPickerPreview() {
-    MaterialTheme {
-        FullColorPicker(
-            uiState = "#FF0000",
             onEvent = {},
             navTo = {}
         )
