+++ settings.gradle.kts
@@ -68,6 +68,7 @@ include(":features:ble")
 include(":features:nfc")
 include(":features:qrscanner")
 include(":features:compliance")
+include(":features:graphics")
 include(":features:ai:capture")
 include(":features:ai:vision")
 include(":features:ai:configuration")
